+++ changelog.md
@@ -15,6 +15,8 @@
   * Add `NFData1` and `NFData2` type classes ([#8](https://github.com/haskell/deepseq/issues/8))
   * Add `NFData1` and `NFData` instances for `Data.Functor.{Compose,Sum,Product}`
     ([#30](https://github.com/haskell/deepseq/pull/30))
+  * Add `NFData`, `NFData1`, and `NFData2` instances for `(:~:)` and `(:~~:)`
+    from `Data.Type.Equality` ([#31](https://github.com/haskell/deepseq/issues/31))
   * Add `NFData` instance for `Foreign.C.Types.CBool`
   * Drop support for GHC 7.0 & GHC 7.2
   * Expose `NFData` instance for `Down` on earlier versions of `base`